refactor(tidy): inline thread purge owner and drop langgraph shells
